Rename the package and cmd
