+++ cmd/lightstep-prometheus-sidecar/main.go
@@ -33,14 +33,14 @@ import (
 	md "cloud.google.com/go/compute/metadata"
 	oc_prometheus "contrib.go.opencensus.io/exporter/prometheus"
 	oc_stackdriver "contrib.go.opencensus.io/exporter/stackdriver"
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/metadata"
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/retrieval"
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/stackdriver"
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/tail"
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/targets"
 	"github.com/ghodss/yaml"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/lightstep/lightstep-prometheus-sidecar/metadata"
+	"github.com/lightstep/lightstep-prometheus-sidecar/retrieval"
+	"github.com/lightstep/lightstep-prometheus-sidecar/stackdriver"
+	"github.com/lightstep/lightstep-prometheus-sidecar/tail"
+	"github.com/lightstep/lightstep-prometheus-sidecar/targets"
 	conntrack "github.com/mwitkow/go-conntrack"
 	"github.com/oklog/oklog/pkg/group"
 	"github.com/pkg/errors"
