+++ cmd/lightstep-prometheus-sidecar/statik/statik.go
