+++ cmd/lightstep-prometheus-sidecar/testdata/wal/000000
