+++ cmd/lightstep-prometheus-sidecar/uname_default.go
