+++ stackdriver/queue_manager_test.go
@@ -24,8 +24,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Stackdriver/stackdriver-prometheus-sidecar/tail"
 	timestamp_pb "github.com/golang/protobuf/ptypes/timestamp"
+	"github.com/lightstep/lightstep-prometheus-sidecar/tail"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/config"
 	metric_pb "google.golang.org/genproto/googleapis/api/metric"
